Batch cherry-picks
- Remove not inuse stale segments in flushmsg (#20981)
- Move APIs of types.DataNode to services.go (#21042)
