+++ internal/datanode/data_node_test.go
@@ -19,36 +19,29 @@ package datanode
 import (
 	"context"
 	"fmt"
-	"math"
 	"math/rand"
 	"os"
-	"path/filepath"
 	"strconv"
 	"strings"
-	"sync"
 	"testing"
 	"time"
 
 	"github.com/golang/protobuf/proto"
+	"github.com/stretchr/testify/assert"
+	"go.uber.org/zap"
+
 	"github.com/milvus-io/milvus-proto/go-api/commonpb"
-	"github.com/milvus-io/milvus-proto/go-api/milvuspb"
 	"github.com/milvus-io/milvus-proto/go-api/schemapb"
-	"github.com/milvus-io/milvus/internal/common"
+
 	etcdkv "github.com/milvus-io/milvus/internal/kv/etcd"
 	"github.com/milvus-io/milvus/internal/log"
-	"github.com/milvus-io/milvus/internal/mq/msgstream"
 	"github.com/milvus-io/milvus/internal/proto/datapb"
-	"github.com/milvus-io/milvus/internal/proto/internalpb"
 	"github.com/milvus-io/milvus/internal/storage"
 	"github.com/milvus-io/milvus/internal/types"
-	"github.com/milvus-io/milvus/internal/util/dependency"
 	"github.com/milvus-io/milvus/internal/util/etcd"
 	"github.com/milvus-io/milvus/internal/util/importutil"
 	"github.com/milvus-io/milvus/internal/util/metricsinfo"
 	"github.com/milvus-io/milvus/internal/util/sessionutil"
-	"github.com/stretchr/testify/assert"
-	"github.com/stretchr/testify/require"
-	"go.uber.org/zap"
 )
 
 const returnError = "ReturnError"
@@ -156,196 +149,6 @@ func TestDataNode(t *testing.T) {
 		}
 	})
 
-	t.Run("Test GetComponentStates", func(t *testing.T) {
-		stat, err := node.GetComponentStates(node.ctx)
-		assert.NoError(t, err)
-		assert.Equal(t, commonpb.ErrorCode_Success, stat.Status.ErrorCode)
-	})
-
-	t.Run("Test GetCompactionState", func(t *testing.T) {
-		node.compactionExecutor.executing.Store(int64(3), 0)
-		node.compactionExecutor.executing.Store(int64(2), 0)
-		node.compactionExecutor.completed.Store(int64(1), &datapb.CompactionResult{
-			PlanID:    1,
-			SegmentID: 10,
-		})
-		stat, err := node.GetCompactionState(node.ctx, nil)
-		assert.NoError(t, err)
-
-		assert.Equal(t, 3, len(stat.GetResults()))
-
-		cnt := 0
-		for _, v := range stat.GetResults() {
-			if v.GetState() == commonpb.CompactionState_Completed {
-				cnt++
-			}
-		}
-		assert.Equal(t, 1, cnt)
-
-		cnt = 0
-		node.compactionExecutor.completed.Range(func(k, v any) bool {
-			cnt++
-			return true
-		})
-		assert.Equal(t, 0, cnt)
-	})
-
-	t.Run("Test GetCompactionState unhealthy", func(t *testing.T) {
-		node.UpdateStateCode(commonpb.StateCode_Abnormal)
-		resp, _ := node.GetCompactionState(ctx, nil)
-		assert.Equal(t, "DataNode is unhealthy", resp.GetStatus().GetReason())
-		node.UpdateStateCode(commonpb.StateCode_Healthy)
-	})
-
-	t.Run("Test FlushSegments", func(t *testing.T) {
-		dmChannelName := "fake-by-dev-rootcoord-dml-channel-test-FlushSegments"
-
-		node1 := newIDLEDataNodeMock(context.TODO(), schemapb.DataType_Int64)
-		node1.SetEtcdClient(etcdCli)
-		err = node1.Init()
-		assert.Nil(t, err)
-		err = node1.Start()
-		assert.Nil(t, err)
-		defer func() {
-			err := node1.Stop()
-			assert.Nil(t, err)
-		}()
-
-		vchan := &datapb.VchannelInfo{
-			CollectionID:        1,
-			ChannelName:         dmChannelName,
-			UnflushedSegmentIds: []int64{},
-			FlushedSegmentIds:   []int64{},
-		}
-
-		err := node1.flowgraphManager.addAndStart(node1, vchan, nil)
-		require.Nil(t, err)
-
-		fgservice, ok := node1.flowgraphManager.getFlowgraphService(dmChannelName)
-		assert.True(t, ok)
-
-		err = fgservice.channel.addSegment(addSegmentReq{
-			segType:     datapb.SegmentType_New,
-			segID:       0,
-			collID:      1,
-			partitionID: 1,
-			startPos:    &internalpb.MsgPosition{},
-			endPos:      &internalpb.MsgPosition{},
-		})
-		assert.Nil(t, err)
-
-		req := &datapb.FlushSegmentsRequest{
-			Base: &commonpb.MsgBase{
-				TargetID: node1.session.ServerID,
-			},
-			DbID:         0,
-			CollectionID: 1,
-			SegmentIDs:   []int64{0},
-		}
-
-		wg := sync.WaitGroup{}
-		wg.Add(2)
-
-		go func() {
-			defer wg.Done()
-
-			status, err := node1.FlushSegments(node1.ctx, req)
-			assert.NoError(t, err)
-			assert.Equal(t, commonpb.ErrorCode_Success, status.ErrorCode)
-		}()
-
-		go func() {
-			defer wg.Done()
-
-			timeTickMsgPack := msgstream.MsgPack{}
-			timeTickMsg := &msgstream.TimeTickMsg{
-				BaseMsg: msgstream.BaseMsg{
-					BeginTimestamp: Timestamp(0),
-					EndTimestamp:   Timestamp(0),
-					HashValues:     []uint32{0},
-				},
-				TimeTickMsg: internalpb.TimeTickMsg{
-					Base: &commonpb.MsgBase{
-						MsgType:   commonpb.MsgType_TimeTick,
-						MsgID:     UniqueID(0),
-						Timestamp: math.MaxUint64,
-						SourceID:  0,
-					},
-				},
-			}
-			timeTickMsgPack.Msgs = append(timeTickMsgPack.Msgs, timeTickMsg)
-
-			// pulsar produce
-			factory := dependency.NewDefaultFactory(true)
-			insertStream, err := factory.NewMsgStream(node1.ctx)
-			assert.NoError(t, err)
-			insertStream.AsProducer([]string{dmChannelName})
-			insertStream.Start()
-			defer insertStream.Close()
-
-			err = insertStream.Broadcast(&timeTickMsgPack)
-			assert.NoError(t, err)
-
-			err = insertStream.Broadcast(&timeTickMsgPack)
-			assert.NoError(t, err)
-		}()
-
-		wg.Wait()
-		// dup call
-		status, err := node1.FlushSegments(node1.ctx, req)
-		assert.NoError(t, err)
-		assert.Equal(t, commonpb.ErrorCode_Success, status.ErrorCode)
-
-		// failure call
-		req = &datapb.FlushSegmentsRequest{
-			Base: &commonpb.MsgBase{
-				TargetID: -1,
-			},
-			DbID:         0,
-			CollectionID: 1,
-			SegmentIDs:   []int64{1},
-		}
-		status, err = node1.FlushSegments(node1.ctx, req)
-		assert.NoError(t, err)
-		assert.Equal(t, commonpb.ErrorCode_NodeIDNotMatch, status.ErrorCode)
-
-		req = &datapb.FlushSegmentsRequest{
-			Base: &commonpb.MsgBase{
-				TargetID: node1.session.ServerID,
-			},
-			DbID:         0,
-			CollectionID: 1,
-			SegmentIDs:   []int64{1},
-		}
-
-		status, err = node1.FlushSegments(node1.ctx, req)
-		assert.NoError(t, err)
-		assert.Equal(t, commonpb.ErrorCode_UnexpectedError, status.ErrorCode)
-
-		req = &datapb.FlushSegmentsRequest{
-			Base: &commonpb.MsgBase{
-				TargetID: node1.session.ServerID,
-			},
-			DbID:         0,
-			CollectionID: 1,
-			SegmentIDs:   []int64{},
-		}
-
-		status, err = node1.FlushSegments(node1.ctx, req)
-		assert.NoError(t, err)
-		assert.Equal(t, commonpb.ErrorCode_Success, status.ErrorCode)
-	})
-
-	t.Run("Test GetTimeTickChannel", func(t *testing.T) {
-		_, err := node.GetTimeTickChannel(node.ctx)
-		assert.NoError(t, err)
-	})
-
-	t.Run("Test GetStatisticsChannel", func(t *testing.T) {
-		_, err := node.GetStatisticsChannel(node.ctx)
-		assert.NoError(t, err)
-	})
-
 	t.Run("Test getSystemInfoMetrics", func(t *testing.T) {
 		emptyNode := &DataNode{}
 		emptyNode.session = &sessionutil.Session{ServerID: 1}
@@ -375,258 +178,6 @@ func TestDataNode(t *testing.T) {
 		rateCol.Register(metricsinfo.InsertConsumeThroughput)
 	})
 
-	t.Run("Test ShowConfigurations", func(t *testing.T) {
-		pattern := "Port"
-		req := &internalpb.ShowConfigurationsRequest{
-			Base: &commonpb.MsgBase{
-				MsgType: commonpb.MsgType_WatchQueryChannels,
-				MsgID:   rand.Int63(),
-			},
-			Pattern: pattern,
-		}
-
-		//test closed server
-		node := &DataNode{}
-		node.session = &sessionutil.Session{ServerID: 1}
-		node.stateCode.Store(commonpb.StateCode_Abnormal)
-
-		resp, err := node.ShowConfigurations(ctx, req)
-		assert.NoError(t, err)
-		assert.Equal(t, commonpb.ErrorCode_UnexpectedError, resp.Status.ErrorCode)
-		node.stateCode.Store(commonpb.StateCode_Healthy)
-
-		resp, err = node.ShowConfigurations(ctx, req)
-		assert.NoError(t, err)
-		assert.Equal(t, commonpb.ErrorCode_Success, resp.Status.ErrorCode)
-
-		assert.Equal(t, 1, len(resp.Configuations))
-		assert.Equal(t, "datanode.port", resp.Configuations[0].Key)
-	})
-
-	t.Run("Test GetMetrics", func(t *testing.T) {
-		node := &DataNode{}
-		node.session = &sessionutil.Session{ServerID: 1}
-		node.flowgraphManager = newFlowgraphManager()
-		// server is closed
-		node.stateCode.Store(commonpb.StateCode_Abnormal)
-		resp, err := node.GetMetrics(ctx, &milvuspb.GetMetricsRequest{})
-		assert.NoError(t, err)
-		assert.NotEqual(t, commonpb.ErrorCode_Success, resp.Status.ErrorCode)
-		node.stateCode.Store(commonpb.StateCode_Healthy)
-
-		// failed to parse metric type
-		invalidRequest := "invalid request"
-		resp, err = node.GetMetrics(ctx, &milvuspb.GetMetricsRequest{
-			Request: invalidRequest,
-		})
-		assert.NoError(t, err)
-		assert.NotEqual(t, commonpb.ErrorCode_Success, resp.Status.ErrorCode)
-
-		// unsupported metric type
-		unsupportedMetricType := "unsupported"
-		req, err := metricsinfo.ConstructRequestByMetricType(unsupportedMetricType)
-		assert.NoError(t, err)
-		resp, err = node.GetMetrics(ctx, req)
-		assert.NoError(t, err)
-		assert.NotEqual(t, commonpb.ErrorCode_Success, resp.Status.ErrorCode)
-
-		// normal case
-		req, err = metricsinfo.ConstructRequestByMetricType(metricsinfo.SystemInfoMetrics)
-		assert.NoError(t, err)
-		resp, err = node.GetMetrics(node.ctx, req)
-		assert.NoError(t, err)
-		assert.Equal(t, commonpb.ErrorCode_Success, resp.Status.ErrorCode)
-		log.Info("Test DataNode.GetMetrics",
-			zap.String("name", resp.ComponentName),
-			zap.String("response", resp.Response))
-	})
-
-	t.Run("Test Import", func(t *testing.T) {
-		node.rootCoord = &RootCoordFactory{
-			collectionID: 100,
-			pkType:       schemapb.DataType_Int64,
-		}
-		content := []byte(`{
-		"rows":[
-			{"bool_field": true, "int8_field": 10, "int16_field": 101, "int32_field": 1001, "int64_field": 10001, "float32_field": 3.14, "float64_field": 1.56, "varChar_field": "hello world", "binary_vector_field": [254, 0, 254, 0], "float_vector_field": [1.1, 1.2]},
-			{"bool_field": false, "int8_field": 11, "int16_field": 102, "int32_field": 1002, "int64_field": 10002, "float32_field": 3.15, "float64_field": 2.56, "varChar_field": "hello world", "binary_vector_field": [253, 0, 253, 0], "float_vector_field": [2.1, 2.2]},
-			{"bool_field": true, "int8_field": 12, "int16_field": 103, "int32_field": 1003, "int64_field": 10003, "float32_field": 3.16, "float64_field": 3.56, "varChar_field": "hello world", "binary_vector_field": [252, 0, 252, 0], "float_vector_field": [3.1, 3.2]},
-			{"bool_field": false, "int8_field": 13, "int16_field": 104, "int32_field": 1004, "int64_field": 10004, "float32_field": 3.17, "float64_field": 4.56, "varChar_field": "hello world", "binary_vector_field": [251, 0, 251, 0], "float_vector_field": [4.1, 4.2]},
-			{"bool_field": true, "int8_field": 14, "int16_field": 105, "int32_field": 1005, "int64_field": 10005, "float32_field": 3.18, "float64_field": 5.56, "varChar_field": "hello world", "binary_vector_field": [250, 0, 250, 0], "float_vector_field": [5.1, 5.2]}
-		]
-		}`)
-
-		chName1 := "fake-by-dev-rootcoord-dml-testimport-1"
-		chName2 := "fake-by-dev-rootcoord-dml-testimport-2"
-		err := node.flowgraphManager.addAndStart(node, &datapb.VchannelInfo{
-			CollectionID:        100,
-			ChannelName:         chName1,
-			UnflushedSegmentIds: []int64{},
-			FlushedSegmentIds:   []int64{},
-		}, nil)
-		require.Nil(t, err)
-		err = node.flowgraphManager.addAndStart(node, &datapb.VchannelInfo{
-			CollectionID:        100,
-			ChannelName:         chName2,
-			UnflushedSegmentIds: []int64{},
-			FlushedSegmentIds:   []int64{},
-		}, nil)
-		require.Nil(t, err)
-
-		_, ok := node.flowgraphManager.getFlowgraphService(chName1)
-		assert.True(t, ok)
-		_, ok = node.flowgraphManager.getFlowgraphService(chName2)
-		assert.True(t, ok)
-
-		filePath := filepath.Join(node.chunkManager.RootPath(), "rows_1.json")
-		err = node.chunkManager.Write(ctx, filePath, content)
-		assert.NoError(t, err)
-		req := &datapb.ImportTaskRequest{
-			ImportTask: &datapb.ImportTask{
-				CollectionId: 100,
-				PartitionId:  100,
-				ChannelNames: []string{chName1, chName2},
-				Files:        []string{filePath},
-				RowBased:     true,
-			},
-		}
-		node.rootCoord.(*RootCoordFactory).ReportImportErr = true
-		_, err = node.Import(node.ctx, req)
-		assert.NoError(t, err)
-		node.rootCoord.(*RootCoordFactory).ReportImportErr = false
-
-		node.rootCoord.(*RootCoordFactory).ReportImportNotSuccess = true
-		_, err = node.Import(context.WithValue(ctx, ctxKey{}, ""), req)
-		assert.NoError(t, err)
-		node.rootCoord.(*RootCoordFactory).ReportImportNotSuccess = false
-
-		node.dataCoord.(*DataCoordFactory).AddSegmentError = true
-		_, err = node.Import(context.WithValue(ctx, ctxKey{}, ""), req)
-		assert.NoError(t, err)
-		node.dataCoord.(*DataCoordFactory).AddSegmentError = false
-
-		node.dataCoord.(*DataCoordFactory).AddSegmentNotSuccess = true
-		_, err = node.Import(context.WithValue(ctx, ctxKey{}, ""), req)
-		assert.NoError(t, err)
-		node.dataCoord.(*DataCoordFactory).AddSegmentNotSuccess = false
-
-		stat, err := node.Import(context.WithValue(ctx, ctxKey{}, ""), req)
-		assert.NoError(t, err)
-		assert.Equal(t, commonpb.ErrorCode_Success, stat.GetErrorCode())
-		assert.Equal(t, "", stat.GetReason())
-	})
-
-	t.Run("Test Import bad flow graph", func(t *testing.T) {
-		node.rootCoord = &RootCoordFactory{
-			collectionID: 100,
-			pkType:       schemapb.DataType_Int64,
-		}
-
-		chName1 := "fake-by-dev-rootcoord-dml-testimport-1"
-		chName2 := "fake-by-dev-rootcoord-dml-testimport-2"
-		err := node.flowgraphManager.addAndStart(node, &datapb.VchannelInfo{
-			CollectionID:        100,
-			ChannelName:         chName1,
-			UnflushedSegmentIds: []int64{},
-			FlushedSegmentIds:   []int64{},
-		}, nil)
-		require.Nil(t, err)
-		err = node.flowgraphManager.addAndStart(node, &datapb.VchannelInfo{
-			CollectionID:        999, // wrong collection ID.
-			ChannelName:         chName2,
-			UnflushedSegmentIds: []int64{},
-			FlushedSegmentIds:   []int64{},
-		}, nil)
-		require.Nil(t, err)
-
-		_, ok := node.flowgraphManager.getFlowgraphService(chName1)
-		assert.True(t, ok)
-		_, ok = node.flowgraphManager.getFlowgraphService(chName2)
-		assert.True(t, ok)
-
-		content := []byte(`{
-		"rows":[
-			{"bool_field": true, "int8_field": 10, "int16_field": 101, "int32_field": 1001, "int64_field": 10001, "float32_field": 3.14, "float64_field": 1.56, "varChar_field": "hello world", "binary_vector_field": [254, 0, 254, 0], "float_vector_field": [1.1, 1.2]},
-			{"bool_field": false, "int8_field": 11, "int16_field": 102, "int32_field": 1002, "int64_field": 10002, "float32_field": 3.15, "float64_field": 2.56, "varChar_field": "hello world", "binary_vector_field": [253, 0, 253, 0], "float_vector_field": [2.1, 2.2]},
-			{"bool_field": true, "int8_field": 12, "int16_field": 103, "int32_field": 1003, "int64_field": 10003, "float32_field": 3.16, "float64_field": 3.56, "varChar_field": "hello world", "binary_vector_field": [252, 0, 252, 0], "float_vector_field": [3.1, 3.2]},
-			{"bool_field": false, "int8_field": 13, "int16_field": 104, "int32_field": 1004, "int64_field": 10004, "float32_field": 3.17, "float64_field": 4.56, "varChar_field": "hello world", "binary_vector_field": [251, 0, 251, 0], "float_vector_field": [4.1, 4.2]},
-			{"bool_field": true, "int8_field": 14, "int16_field": 105, "int32_field": 1005, "int64_field": 10005, "float32_field": 3.18, "float64_field": 5.56, "varChar_field": "hello world", "binary_vector_field": [250, 0, 250, 0], "float_vector_field": [5.1, 5.2]}
-		]
-		}`)
-
-		filePath := filepath.Join(node.chunkManager.RootPath(), "rows_1.json")
-		err = node.chunkManager.Write(ctx, filePath, content)
-		assert.NoError(t, err)
-		req := &datapb.ImportTaskRequest{
-			ImportTask: &datapb.ImportTask{
-				CollectionId: 100,
-				PartitionId:  100,
-				ChannelNames: []string{chName1, chName2},
-				Files:        []string{filePath},
-				RowBased:     true,
-			},
-		}
-		stat, err := node.Import(context.WithValue(ctx, ctxKey{}, ""), req)
-		assert.NoError(t, err)
-		assert.Equal(t, commonpb.ErrorCode_Success, stat.GetErrorCode())
-		assert.Equal(t, "", stat.GetReason())
-	})
-
-	t.Run("Test Import report import error", func(t *testing.T) {
-		node.rootCoord = &RootCoordFactory{
-			collectionID:    100,
-			pkType:          schemapb.DataType_Int64,
-			ReportImportErr: true,
-		}
-		content := []byte(`{
-		"rows":[
-			{"bool_field": true, "int8_field": 10, "int16_field": 101, "int32_field": 1001, "int64_field": 10001, "float32_field": 3.14, "float64_field": 1.56, "varChar_field": "hello world", "binary_vector_field": [254, 0, 254, 0], "float_vector_field": [1.1, 1.2]},
-			{"bool_field": false, "int8_field": 11, "int16_field": 102, "int32_field": 1002, "int64_field": 10002, "float32_field": 3.15, "float64_field": 2.56, "varChar_field": "hello world", "binary_vector_field": [253, 0, 253, 0], "float_vector_field": [2.1, 2.2]},
-			{"bool_field": true, "int8_field": 12, "int16_field": 103, "int32_field": 1003, "int64_field": 10003, "float32_field": 3.16, "float64_field": 3.56, "varChar_field": "hello world", "binary_vector_field": [252, 0, 252, 0], "float_vector_field": [3.1, 3.2]},
-			{"bool_field": false, "int8_field": 13, "int16_field": 104, "int32_field": 1004, "int64_field": 10004, "float32_field": 3.17, "float64_field": 4.56, "varChar_field": "hello world", "binary_vector_field": [251, 0, 251, 0], "float_vector_field": [4.1, 4.2]},
-			{"bool_field": true, "int8_field": 14, "int16_field": 105, "int32_field": 1005, "int64_field": 10005, "float32_field": 3.18, "float64_field": 5.56, "varChar_field": "hello world", "binary_vector_field": [250, 0, 250, 0], "float_vector_field": [5.1, 5.2]}
-		]
-		}`)
-
-		filePath := filepath.Join(node.chunkManager.RootPath(), "rows_1.json")
-		err = node.chunkManager.Write(ctx, filePath, content)
-		assert.NoError(t, err)
-		req := &datapb.ImportTaskRequest{
-			ImportTask: &datapb.ImportTask{
-				CollectionId: 100,
-				PartitionId:  100,
-				ChannelNames: []string{"ch1", "ch2"},
-				Files:        []string{filePath},
-				RowBased:     true,
-			},
-		}
-		stat, err := node.Import(node.ctx, req)
-		assert.NoError(t, err)
-		assert.Equal(t, commonpb.ErrorCode_UnexpectedError, stat.GetErrorCode())
-	})
-
-	t.Run("Test Import error", func(t *testing.T) {
-		node.rootCoord = &RootCoordFactory{collectionID: -1}
-		req := &datapb.ImportTaskRequest{
-			ImportTask: &datapb.ImportTask{
-				CollectionId: 100,
-				PartitionId:  100,
-			},
-		}
-		stat, err := node.Import(context.WithValue(ctx, ctxKey{}, ""), req)
-		assert.NoError(t, err)
-		assert.Equal(t, commonpb.ErrorCode_UnexpectedError, stat.ErrorCode)
-
-		stat, err = node.Import(context.WithValue(ctx, ctxKey{}, returnError), req)
-		assert.NoError(t, err)
-		assert.Equal(t, commonpb.ErrorCode_UnexpectedError, stat.GetErrorCode())
-
-		node.stateCode.Store(commonpb.StateCode_Abnormal)
-		stat, err = node.Import(context.WithValue(ctx, ctxKey{}, ""), req)
-		assert.NoError(t, err)
-		assert.Equal(t, commonpb.ErrorCode_UnexpectedError, stat.GetErrorCode())
-	})
-
 	t.Run("Test BackGroundGC", func(t *testing.T) {
 		ctx, cancel := context.WithCancel(context.Background())
 		node := newIDLEDataNodeMock(ctx, schemapb.DataType_Int64)
@@ -653,184 +204,6 @@ func TestDataNode(t *testing.T) {
 			}
 		}
 		cancel()
-	})
-
-	t.Run("Test SyncSegments", func(t *testing.T) {
-		ctx, cancel := context.WithCancel(context.Background())
-		defer cancel()
-		chanName := "fake-by-dev-rootcoord-dml-test-syncsegments-1"
-
-		node := newIDLEDataNodeMock(ctx, schemapb.DataType_Int64)
-		etcdCli, err := etcd.GetEtcdClient(
-			Params.EtcdCfg.UseEmbedEtcd,
-			Params.EtcdCfg.EtcdUseSSL,
-			Params.EtcdCfg.Endpoints,
-			Params.EtcdCfg.EtcdTLSCert,
-			Params.EtcdCfg.EtcdTLSKey,
-			Params.EtcdCfg.EtcdTLSCACert,
-			Params.EtcdCfg.EtcdTLSMinVersion)
-		assert.Nil(t, err)
-		defer etcdCli.Close()
-		node.SetEtcdClient(etcdCli)
-		err = node.Init()
-		assert.Nil(t, err)
-		err = node.Start()
-		assert.Nil(t, err)
-		defer node.Stop()
-
-		err = node.flowgraphManager.addAndStart(node, &datapb.VchannelInfo{
-			ChannelName:         chanName,
-			UnflushedSegmentIds: []int64{},
-			FlushedSegmentIds:   []int64{},
-		}, nil)
-		require.NoError(t, err)
-		fg, ok := node.flowgraphManager.getFlowgraphService(chanName)
-		assert.True(t, ok)
-
-		s1 := Segment{segmentID: 100}
-		s2 := Segment{segmentID: 200}
-		s3 := Segment{segmentID: 300}
-		s1.setType(datapb.SegmentType_Flushed)
-		s2.setType(datapb.SegmentType_Flushed)
-		s3.setType(datapb.SegmentType_Flushed)
-		fg.channel.(*ChannelMeta).segments = map[UniqueID]*Segment{
-			s1.segmentID: &s1,
-			s2.segmentID: &s2,
-			s3.segmentID: &s3,
-		}
-
-		t.Run("invalid compacted from", func(t *testing.T) {
-			invalidCompactedFroms := [][]UniqueID{
-				{},
-				{101, 201},
-			}
-			req := &datapb.SyncSegmentsRequest{}
-
-			for _, invalid := range invalidCompactedFroms {
-				req.CompactedFrom = invalid
-				status, err := node.SyncSegments(ctx, req)
-				assert.NoError(t, err)
-				assert.Equal(t, commonpb.ErrorCode_UnexpectedError, status.GetErrorCode())
-			}
-		})
-
-		t.Run("valid request numRows>0", func(t *testing.T) {
-			req := &datapb.SyncSegmentsRequest{
-				CompactedFrom: []int64{100, 200},
-				CompactedTo:   101,
-				NumOfRows:     100,
-			}
-			status, err := node.SyncSegments(ctx, req)
-			assert.NoError(t, err)
-			assert.Equal(t, commonpb.ErrorCode_Success, status.GetErrorCode())
-
-			assert.True(t, fg.channel.hasSegment(req.CompactedTo, true))
-			assert.False(t, fg.channel.hasSegment(req.CompactedFrom[0], true))
-			assert.False(t, fg.channel.hasSegment(req.CompactedFrom[1], true))
-		})
-
-		t.Run("valid request numRows=0", func(t *testing.T) {
-			s1.setType(datapb.SegmentType_Flushed)
-			s2.setType(datapb.SegmentType_Flushed)
-			s3.setType(datapb.SegmentType_Flushed)
-
-			fg.channel.(*ChannelMeta).segments = map[UniqueID]*Segment{
-				s1.segmentID: &s1,
-				s2.segmentID: &s2,
-				s3.segmentID: &s3,
-			}
-
-			req := &datapb.SyncSegmentsRequest{
-				CompactedFrom: []int64{s1.segmentID, s2.segmentID},
-				CompactedTo:   101,
-				NumOfRows:     0,
-			}
-			status, err := node.SyncSegments(ctx, req)
-			assert.NoError(t, err)
-			assert.Equal(t, commonpb.ErrorCode_Success, status.GetErrorCode())
-
-			assert.False(t, fg.channel.hasSegment(req.CompactedTo, true))
-			assert.False(t, fg.channel.hasSegment(req.CompactedFrom[0], true))
-			assert.False(t, fg.channel.hasSegment(req.CompactedFrom[1], true))
-		})
-	})
-}
-
-func TestDataNode_AddSegment(t *testing.T) {
-	ctx, cancel := context.WithCancel(context.Background())
-	defer cancel()
-
-	node := newIDLEDataNodeMock(ctx, schemapb.DataType_Int64)
-	etcdCli, err := etcd.GetEtcdClient(
-		Params.EtcdCfg.UseEmbedEtcd,
-		Params.EtcdCfg.EtcdUseSSL,
-		Params.EtcdCfg.Endpoints,
-		Params.EtcdCfg.EtcdTLSCert,
-		Params.EtcdCfg.EtcdTLSKey,
-		Params.EtcdCfg.EtcdTLSCACert,
-		Params.EtcdCfg.EtcdTLSMinVersion)
-	assert.Nil(t, err)
-	defer etcdCli.Close()
-	node.SetEtcdClient(etcdCli)
-	err = node.Init()
-	assert.Nil(t, err)
-	err = node.Start()
-	assert.Nil(t, err)
-	defer node.Stop()
-
-	node.chunkManager = storage.NewLocalChunkManager(storage.RootPath("/tmp/milvus_test/datanode"))
-	Params.DataNodeCfg.SetNodeID(1)
-
-	t.Run("test AddSegment", func(t *testing.T) {
-		node.rootCoord = &RootCoordFactory{
-			collectionID: 100,
-			pkType:       schemapb.DataType_Int64,
-		}
-
-		chName1 := "fake-by-dev-rootcoord-dml-testaddsegment-1"
-		chName2 := "fake-by-dev-rootcoord-dml-testaddsegment-2"
-		err := node.flowgraphManager.addAndStart(node, &datapb.VchannelInfo{
-			CollectionID:        100,
-			ChannelName:         chName1,
-			UnflushedSegmentIds: []int64{},
-			FlushedSegmentIds:   []int64{},
-		}, nil)
-		require.Nil(t, err)
-		err = node.flowgraphManager.addAndStart(node, &datapb.VchannelInfo{
-			CollectionID:        100,
-			ChannelName:         chName2,
-			UnflushedSegmentIds: []int64{},
-			FlushedSegmentIds:   []int64{},
-		}, nil)
-		require.Nil(t, err)
-
-		_, ok := node.flowgraphManager.getFlowgraphService(chName1)
-		assert.True(t, ok)
-		_, ok = node.flowgraphManager.getFlowgraphService(chName2)
-		assert.True(t, ok)
-
-		stat, err := node.AddImportSegment(context.WithValue(ctx, ctxKey{}, ""), &datapb.AddImportSegmentRequest{
-			SegmentId:    100,
-			CollectionId: 100,
-			PartitionId:  100,
-			ChannelName:  chName1,
-			RowNum:       500,
-		})
-		assert.NoError(t, err)
-		assert.Equal(t, commonpb.ErrorCode_Success, stat.GetStatus().GetErrorCode())
-		assert.Equal(t, "", stat.GetStatus().GetReason())
-		assert.NotEqual(t, nil, stat.GetChannelPos())
-
-		getFlowGraphServiceAttempts = 3
-		stat, err = node.AddImportSegment(context.WithValue(ctx, ctxKey{}, ""), &datapb.AddImportSegmentRequest{
-			SegmentId:    100,
-			CollectionId: 100,
-			PartitionId:  100,
-			ChannelName:  "bad-ch-name",
-			RowNum:       500,
-		})
-		assert.NoError(t, err)
-		assert.Equal(t, commonpb.ErrorCode_UnexpectedError, stat.GetStatus().GetErrorCode())
 	})
 }
 
@@ -1098,102 +471,4 @@ func TestWatchChannel(t *testing.T) {
 		assert.NotEmpty(t, newWatchInfo.GetVchan().GetFlushedSegmentIds())
 		assert.NotEmpty(t, newWatchInfo.GetVchan().GetDroppedSegmentIds())
 	})
-}
-
-func TestDataNode_GetComponentStates(t *testing.T) {
-	n := &DataNode{}
-	n.stateCode.Store(commonpb.StateCode_Healthy)
-	resp, err := n.GetComponentStates(context.Background())
-	assert.NoError(t, err)
-	assert.Equal(t, commonpb.ErrorCode_Success, resp.Status.ErrorCode)
-	assert.Equal(t, common.NotRegisteredID, resp.State.NodeID)
-	n.session = &sessionutil.Session{}
-	n.session.UpdateRegistered(true)
-	resp, err = n.GetComponentStates(context.Background())
-	assert.NoError(t, err)
-	assert.Equal(t, commonpb.ErrorCode_Success, resp.Status.ErrorCode)
-}
-
-func TestDataNode_ResendSegmentStats(t *testing.T) {
-	etcdCli, err := etcd.GetEtcdClient(
-		Params.EtcdCfg.UseEmbedEtcd,
-		Params.EtcdCfg.EtcdUseSSL,
-		Params.EtcdCfg.Endpoints,
-		Params.EtcdCfg.EtcdTLSCert,
-		Params.EtcdCfg.EtcdTLSKey,
-		Params.EtcdCfg.EtcdTLSCACert,
-		Params.EtcdCfg.EtcdTLSMinVersion)
-	assert.Nil(t, err)
-	defer etcdCli.Close()
-	dmChannelName := "fake-by-dev-rootcoord-dml-channel-test-ResendSegmentStats"
-
-	node := newIDLEDataNodeMock(context.TODO(), schemapb.DataType_Int64)
-	node.SetEtcdClient(etcdCli)
-	err = node.Init()
-	assert.Nil(t, err)
-	err = node.Start()
-	assert.Nil(t, err)
-	defer func() {
-		err := node.Stop()
-		assert.Nil(t, err)
-	}()
-
-	vChan := &datapb.VchannelInfo{
-		CollectionID:        1,
-		ChannelName:         dmChannelName,
-		UnflushedSegmentIds: []int64{},
-		FlushedSegmentIds:   []int64{},
-	}
-
-	err = node.flowgraphManager.addAndStart(node, vChan, nil)
-	require.Nil(t, err)
-
-	fgService, ok := node.flowgraphManager.getFlowgraphService(dmChannelName)
-	assert.True(t, ok)
-
-	err = fgService.channel.addSegment(addSegmentReq{
-		segType:     datapb.SegmentType_New,
-		segID:       0,
-		collID:      1,
-		partitionID: 1,
-		startPos:    &internalpb.MsgPosition{},
-		endPos:      &internalpb.MsgPosition{},
-	})
-	assert.Nil(t, err)
-	err = fgService.channel.addSegment(addSegmentReq{
-		segType:     datapb.SegmentType_New,
-		segID:       1,
-		collID:      1,
-		partitionID: 2,
-		startPos:    &internalpb.MsgPosition{},
-		endPos:      &internalpb.MsgPosition{},
-	})
-	assert.Nil(t, err)
-	err = fgService.channel.addSegment(addSegmentReq{
-		segType:     datapb.SegmentType_New,
-		segID:       2,
-		collID:      1,
-		partitionID: 3,
-		startPos:    &internalpb.MsgPosition{},
-		endPos:      &internalpb.MsgPosition{},
-	})
-	assert.Nil(t, err)
-
-	req := &datapb.ResendSegmentStatsRequest{
-		Base: &commonpb.MsgBase{},
-	}
-
-	wg := sync.WaitGroup{}
-	wg.Add(2)
-
-	resp, err := node.ResendSegmentStats(node.ctx, req)
-	assert.NoError(t, err)
-	assert.Equal(t, commonpb.ErrorCode_Success, resp.GetStatus().GetErrorCode())
-	assert.ElementsMatch(t, []UniqueID{0, 1, 2}, resp.GetSegResent())
-
-	// Duplicate call.
-	resp, err = node.ResendSegmentStats(node.ctx, req)
-	assert.NoError(t, err)
-	assert.Equal(t, commonpb.ErrorCode_Success, resp.GetStatus().GetErrorCode())
-	assert.ElementsMatch(t, []UniqueID{0, 1, 2}, resp.GetSegResent())
 }
