+++ internal/datanode/flow_graph_insert_buffer_node_test.go
@@ -515,10 +515,7 @@ func TestFlowGraphInsertBufferNode_AutoFlush(t *testing.T) {
 		inMsg.endPositions = []*internalpb.MsgPosition{{Timestamp: 434}}
 
 		// trigger manual flush
-		flushChan <- flushMsg{
-			segmentID: 10,
-			flushed:   true,
-		}
+		flushChan <- flushMsg{segmentID: 10}
 
 		// trigger auto flush since buffer full
 		output := iBNode.Operate([]flowgraph.Msg{iMsg})
@@ -673,7 +670,7 @@ func TestRollBF(t *testing.T) {
 	})
 }
 
-type InsertBufferNodeSuit struct {
+type InsertBufferNodeSuite struct {
 	suite.Suite
 
 	channel *ChannelMeta
@@ -685,7 +682,7 @@ type InsertBufferNodeSuit struct {
 	originalConfig int64
 }
 
-func (s *InsertBufferNodeSuit) SetupSuite() {
+func (s *InsertBufferNodeSuite) SetupSuite() {
 	insertBufferNodeTestDir := "/tmp/milvus_test/insert_buffer_node"
 	rc := &RootCoordFactory{
 		pkType: schemapb.DataType_Int64,
@@ -701,12 +698,12 @@ func (s *InsertBufferNodeSuit) SetupSuite() {
 	Params.DataNodeCfg.FlushInsertBufferSize = 200
 }
 
-func (s *InsertBufferNodeSuit) TearDownSuite() {
+func (s *InsertBufferNodeSuite) TearDownSuite() {
 	s.cm.RemoveWithPrefix(context.Background(), s.cm.RootPath())
 	Params.DataNodeCfg.FlushInsertBufferSize = s.originalConfig
 }
 
-func (s *InsertBufferNodeSuit) SetupTest() {
+func (s *InsertBufferNodeSuite) SetupTest() {
 	segs := []struct {
 		segID UniqueID
 		sType datapb.SegmentType
@@ -729,11 +726,11 @@ func (s *InsertBufferNodeSuit) SetupTest() {
 	}
 }
 
-func (s *InsertBufferNodeSuit) TearDownTest() {
+func (s *InsertBufferNodeSuite) TearDownTest() {
 	s.channel.removeSegments(1, 2, 3)
 }
 
-func (s *InsertBufferNodeSuit) TestFillInSyncTasks() {
+func (s *InsertBufferNodeSuite) TestFillInSyncTasks() {
 	s.Run("drop collection", func() {
 		fgMsg := &flowGraphMsg{dropCollection: true}
 
@@ -812,7 +809,6 @@ func (s *InsertBufferNodeSuit) TestFillInSyncTasks() {
 			msg := flushMsg{
 				segmentID:    UniqueID(i),
 				collectionID: s.collID,
-				flushed:      i%2 == 0, // segID=2, flushed = true
 			}
 
 			flushCh <- msg
@@ -821,13 +817,8 @@ func (s *InsertBufferNodeSuit) TestFillInSyncTasks() {
 		syncTasks := node.FillInSyncTasks(&flowGraphMsg{endPositions: []*internalpb.MsgPosition{{Timestamp: 100}}}, nil)
 		s.Assert().NotEmpty(syncTasks)
 
-		for segID, task := range syncTasks {
-			if segID == UniqueID(2) {
-				s.Assert().True(task.flushed)
-			} else {
-				s.Assert().False(task.flushed)
-			}
-
+		for _, task := range syncTasks {
+			s.Assert().True(task.flushed)
 			s.Assert().False(task.auto)
 			s.Assert().False(task.dropped)
 		}
@@ -845,11 +836,6 @@ func (s *InsertBufferNodeSuit) TestFillInSyncTasks() {
 			msg := flushMsg{
 				segmentID:    UniqueID(i),
 				collectionID: s.collID,
-				flushed:      false,
-			}
-
-			if i == 2 {
-				msg.flushed = true
 			}
 
 			flushCh <- msg
@@ -859,13 +845,8 @@ func (s *InsertBufferNodeSuit) TestFillInSyncTasks() {
 		s.Assert().NotEmpty(syncTasks)
 		s.Assert().Equal(10, len(syncTasks)) // 10 is max batch
 
-		for segID, task := range syncTasks {
-			if segID == UniqueID(2) {
-				s.Assert().True(task.flushed)
-			} else {
-				s.Assert().False(task.flushed)
-			}
-
+		for _, task := range syncTasks {
+			s.Assert().True(task.flushed)
 			s.Assert().False(task.auto)
 			s.Assert().False(task.dropped)
 		}
@@ -874,7 +855,7 @@ func (s *InsertBufferNodeSuit) TestFillInSyncTasks() {
 }
 
 func TestInsertBufferNodeSuite(t *testing.T) {
-	suite.Run(t, new(InsertBufferNodeSuit))
+	suite.Run(t, new(InsertBufferNodeSuite))
 }
 
 // CompactedRootCoord has meta info compacted at ts
@@ -1074,14 +1055,11 @@ func TestInsertBufferNode_collectSegmentsToSync(t *testing.T) {
 			}
 
 			for i := 0; i < test.inFlushMsgNum; i++ {
-				flushCh <- flushMsg{
-					segmentID: UniqueID(i),
-					flushed:   i%2 == 0,
-				}
+				flushCh <- flushMsg{segmentID: UniqueID(i)}
 			}
 
-			flushedSegs, staleSegs := node.CollectSegmentsToSync()
-			assert.Equal(t, test.expectedOutNum, len(flushedSegs)+len(staleSegs))
+			flushedSegs := node.CollectSegmentsToSync()
+			assert.Equal(t, test.expectedOutNum, len(flushedSegs))
 		})
 	}
 }
