+++ internal/datanode/flow_graph_message.go
@@ -62,7 +62,6 @@ type flushMsg struct {
 	timestamp    Timestamp
 	segmentID    UniqueID
 	collectionID UniqueID
-	flushed      bool
 }
 
 type resendTTMsg struct {
